explicit bigint conversion for workflow_timeout_ms update during dequeue

+++ dbos/system_database.go
@@ -2907,7 +2907,7 @@ func (s *sysDB) dequeueWorkflows(ctx context.Context, input dequeueWorkflowsInpu
 			    started_at_epoch_ms = $4,
 			    workflow_deadline_epoch_ms = CASE
 			        WHEN workflow_timeout_ms IS NOT NULL AND workflow_deadline_epoch_ms IS NULL
-			        THEN EXTRACT(epoch FROM NOW()) * 1000 + workflow_timeout_ms
+			        THEN (EXTRACT(epoch FROM NOW()) * 1000)::BIGINT + workflow_timeout_ms
 			        ELSE workflow_deadline_epoch_ms
 			    END
 			WHERE workflow_uuid = $5
